Conection to mongodb through environment variables
