+++ DataRetrieval/build.sbt
@@ -9,6 +9,5 @@ libraryDependencies ++= Seq(
 	"org.apache.spark" % "spark-core_2.11" % "2.0.0",
 	"org.apache.spark" % "spark-sql_2.11" % "2.0.0",
 	"edu.ucar" % "netcdf" % "4.2.20",
-	"org.mongodb.scala" %% "mongo-scala-driver" % "2.4.2",
 	"org.mongodb.spark" % "mongo-spark-connector_2.11" % "2.3.1"
 )
