--- DataRetrieval/src/main/scala/Util/ArgoFloatException.scala
@@ -1,4 +1,0 @@
-package Util
-
-final class ArgoFloatException( private val message: String = "Error trying to create an ArgoFloatObject from raw data",
-                                private val cause: Throwable = None.orNull) extends Exception(message, cause)
